+++ src/nautilus-view-frame.h
@@ -115,11 +115,11 @@ void               nautilus_view_frame_title_changed             (NautilusViewFr
                                                                   const char          *title);
 
 /* calls to Bonobo:Zoomable functions */
-double             nautilus_view_frame_get_zoom_level            (NautilusViewFrame   *view);
+float              nautilus_view_frame_get_zoom_level            (NautilusViewFrame   *view);
 void               nautilus_view_frame_set_zoom_level            (NautilusViewFrame   *view,
-                                                                  double               zoom_level);
-double             nautilus_view_frame_get_min_zoom_level        (NautilusViewFrame   *view);
-double             nautilus_view_frame_get_max_zoom_level        (NautilusViewFrame   *view);
+                                                                  float                zoom_level);
+float              nautilus_view_frame_get_min_zoom_level        (NautilusViewFrame   *view);
+float              nautilus_view_frame_get_max_zoom_level        (NautilusViewFrame   *view);
 gboolean           nautilus_view_frame_get_has_min_zoom_level    (NautilusViewFrame   *view);
 gboolean           nautilus_view_frame_get_has_max_zoom_level    (NautilusViewFrame   *view);
 gboolean           nautilus_view_frame_get_is_continuous         (NautilusViewFrame   *view);
